+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - Offer details controller will now show a dismiss button when presented modally
+- `WakupManager` now provides access to the offer details controller
 
 ## [2.0.0] - 2016-10-26
 ### Added